# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /dailyStatus/year/{year}/month/{month}/day/{day}:
    # binds a127 app logic to a route
    x-swagger-router-controller: dailyStatusController
    get:
      tags: 
        - DailyStatus
      description: Get daily status by date
      # used as the method name of the controller
      operationId: getDailyStatusByDate
      parameters:
        - name: year
          in: path
          description: The year of the daily status report
          required: true
          type: string
        - name: month
          in: path
          description: The month of the daily status report
          required: true
          type: string
        - name: day
          in: path
          description: The day of the daily status report
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/DailyStatus"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /dailyStatus:
    x-swagger-router-controller: dailyStatusController
    post:
      tags: 
        - DailyStatus
      description: Add user daily status
      operationId: addUserDailyStatus
      parameters:
        - $ref: "#/parameters/AddOrUpdateDailyStatusBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags: 
        - DailyStatus
      description: Update user daily status
      operationId: updateUserDailyStatus
      parameters:
        - $ref: "#/parameters/AddOrUpdateDailyStatusBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /dailyStatus/engineer/{engineer}/year/{year}/month/{month}/day/{day}:
    # binds a127 app logic to a route
    x-swagger-router-controller: dailyStatusController
    get:
      tags: 
        - DailyStatus
      description: Get user daily status
      # used as the method name of the controller
      operationId: getUserDailyStatus
      parameters:
        - name: engineer
          in: path
          description: The engineer of the daily status report
          required: true
          type: string
        - name: year
          in: path
          description: The year of the daily status report
          required: true
          type: string
        - name: month
          in: path
          description: The month of the daily status report
          required: true
          type: string
        - name: day
          in: path
          description: The day of the daily status report
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/DailyStatus"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /user/login:
    # binds a127 app logic to a route
    x-swagger-router-controller: userController
    post:
      tags: 
        - User
      description: User login
      # used as the method name of the controller
      operationId: login
      parameters:
        - $ref: "#/parameters/UserLoginBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResult"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /user/register:
    x-swagger-router-controller: userController
    post:
      tags: 
        - User
      description: User register
      operationId: register
      parameters:
        - $ref: "#/parameters/UserLoginBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /user/changePassword:
    x-swagger-router-controller: userController
    post:
      tags: 
        - User
      description: Change user password
      operationId: changePassword
      parameters:
        - $ref: "#/parameters/UserLoginBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /user/list:
    x-swagger-router-controller: userController
    get:
      tags: 
        - User
      description: Get user list
      operationId: getUserList
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /mail/sendMail:
    # binds a127 app logic to a route
    x-swagger-router-controller: mailController
    post:
      tags: 
        - Mail
      description: Send mail
      # used as the method name of the controller
      operationId: sendMail
      parameters:
        - $ref: "#/parameters/SendMailBody"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ErrorResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  DailyStatus:
    required:
      - engineer
      - workOn
      - workItem
      - nextWorkItem
    properties:
      engineer:
        type: string
      workOn:
        type: string
      workItem:
        type: string
      nextWorkItem:
        type: string
      year:
        type: string
      month:
        type: string
      day:
        type: string
  User:
    required:
      - userName
      - password
    properties:
      userName:
        type: string
      password:
        type: string
  Mail:
    required:
      - content
    properties:
      content:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  LoginResult:
    required:
      - userName
    properties:
      userName:
        type: string
parameters:
  AddOrUpdateDailyStatusBody:
    name: body
    description: The daily status to be added to local db.
    in: body
    required: true
    schema:
      $ref: "#/definitions/DailyStatus"
  UserLoginBody:
    name: body
    description: The user login body.
    in: body
    required: true
    schema:
      $ref: "#/definitions/User"
  SendMailBody:
    name: body
    description: The send mail body.
    in: body
    required: true
    schema:
      $ref: "#/definitions/Mail"